+++ packages/dock-blockchain-modules/src/accumulator/index.js
@@ -1,3 +1,0 @@
-export * from './actions';
-export { default as DockAccumulatorModuleInternal } from './internal';
-export { default as DockAccumulatorModule } from './module';