+++ packages/dock-blockchain-modules/src/blob/index.js
@@ -1,4 +1,0 @@
-export * from './actions';
-export { default as DockBlobModuleInternal } from './internal';
-export { default as DockBlobModule } from './module';
-export * from './types';