+++ packages/dock-blockchain-modules/src/common/index.js
@@ -1,7 +1,0 @@
-export { default as createInternalDockModule } from './create-internal-dock-module';
-export { default as injectDock } from './inject-dock';
-export { default as withParams } from './with-params';
-export { default as withPublicKeys } from './with-public-keys';
-export { default as injectParams } from './inject-params';
-export { default as injectPublicKeys } from './inject-public-keys';
-export * from './keypair';