+++ packages/dock-blockchain-modules/src/status-list-credential/index.js
@@ -1,4 +1,0 @@
-export * from './actions';
-export { default as DockInternalStatusListCredentialModule } from './internal';
-export { default as DockStatusListCredentialModule } from './module';
-export * from './types';