+++ scripts/bench/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @docknetwork/benchmarks
 
+## 0.2.0
+
+### Minor Changes
+
+- Package cleanup
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/cheqd-blockchain-modules@1.0.0
+  - @docknetwork/cheqd-blockchain-api@1.0.0
+  - @docknetwork/credential-sdk@0.51.0
+
 ## 0.1.5
 
 ### Patch Changes